Move js parsers to a separate directory
