+++ .eslintrc.js
@@ -217,11 +217,11 @@ module.exports = {
             file: "src/language-js/utils.js",
             functions: ["hasComment", "getComments"],
           },
-          "src/language-js/parse-postprocess.js",
-          "src/language-js/parser-babel.js",
-          "src/language-js/parser-meriyah.js",
           "src/language-js/pragma.js",
-          "src/language-js/parser/json.js",
+          "src/language-js/parse/postprocess.js",
+          "src/language-js/parse/babel.js",
+          "src/language-js/parse/meriyah.js",
+          "src/language-js/parse/json.js",
         ],
       },
     },
