+++ src/language-js/parse/angular.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const { locStart, locEnd } = require("./loc");
+const { locStart, locEnd } = require("../loc");
 
 function createParser(_parse) {
   const parse = (text, parsers, options) => {
