+++ src/language-js/parse/replace-hashbang.js
